+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # TableTest Plugin Changelog
 
 ## [Unreleased]
+
+## [0.1.0] - 2026-01-16
 ### Added
 - Move row up/down with keyboard shortcuts (Cmd+Shift+Up/Down on Mac, Alt+Shift+Up/Down on Windows/Linux)
 ### Fixed